# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple Go API with optional PostgreSQL, Redis, and MongoDB integrations",
        "title": "Simple Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/mongo/documents": {
            "get": {
                "description": "Get all documents from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mongo"
                ],
                "summary": "Get all MongoDB documents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/mongo.Document"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new document in MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mongo"
                ],
                "summary": "Create MongoDB document",
                "parameters": [
                    {
                        "description": "Document data",
                        "name": "document",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.DocumentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mongo.Document"
                        }
                    }
                }
            }
        },
        "/mongo/documents/{id}": {
            "get": {
                "description": "Get a document from MongoDB by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mongo"
                ],
                "summary": "Get MongoDB document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.Document"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a document in MongoDB by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mongo"
                ],
                "summary": "Update MongoDB document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document data",
                        "name": "document",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mongo.DocumentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mongo.Document"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a document from MongoDB by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mongo"
                ],
                "summary": "Delete MongoDB document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mongo.Document": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "mongo.DocumentInput": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}